+++ execution/builtins/env.c
@@ -6,7 +6,7 @@
 /*   By: noaziki <noaziki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/05 17:25:04 by noaziki           #+#    #+#             */
-/*   Updated: 2025/07/03 21:02:12 by noaziki          ###   ########.fr       */
+/*   Updated: 2025/07/04 09:26:16 by noaziki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,10 +32,8 @@ int	env(t_env *env_list, t_stash *stash, char **cmd)
 			if (stash->path_flag == 1 && !ft_strcmp(env_list->key, "PATH"))
 				env_list = env_list->next;
 			if (env_list)
-			{
 				if (env_list->value || !ft_strcmp(env_list->value, ""))
 					printf("%s=%s\n", env_list->key, env_list->value);
-			}
 			if (env_list)
 				env_list = env_list->next;
 		}
