+++ execution/builtins/path_utils.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   cd_utils.c                                         :+:      :+:    :+:   */
+/*   path_utils.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: noaziki <noaziki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/07/02 15:05:54 by noaziki           #+#    #+#             */
-/*   Updated: 2025/07/03 16:06:17 by noaziki          ###   ########.fr       */
+/*   Created: 2025/07/04 09:10:55 by noaziki           #+#    #+#             */
+/*   Updated: 2025/07/04 09:11:55 by noaziki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,19 +51,10 @@ char	*get_next_component(const char *path, int *start, int *end)
 	return (na_substr(path, *start, *end - *start));
 }
 
-char	*build_target_path(int dotdots, const char *component, t_stash *stash)
+char	*handle_dotdots(char *new_path, int dotdots)
 {
-	char	*pwd;
-	char	*new_path;
 	char	*temp;
-	char	*temp2;
 
-	if (!stash || !stash->pwd_backup)
-		return (NULL);
-	pwd = stash->pwd_backup;
-	new_path = na_strdup(pwd);
-	if (!new_path)
-		return (NULL);
 	while (dotdots-- > 0)
 	{
 		temp = ft_strrchr(new_path, '/');
@@ -71,8 +62,16 @@ char	*build_target_path(int dotdots, const char *component, t_stash *stash)
 			return (NULL);
 		*temp = '\0';
 		if (temp == new_path)
-			break;
+			break ;
 	}
+	return (new_path);
+}
+
+char	*handle_component(char *new_path, const char *component)
+{
+	char	*temp;
+	char	*temp2;
+
 	if (component && *component)
 	{
 		temp = na_strjoin(new_path, "/");
@@ -84,49 +83,4 @@ char	*build_target_path(int dotdots, const char *component, t_stash *stash)
 			return (NULL);
 	}
 	return (new_path);
-}
-
-char	*process_cd_path(const char *path, t_stash *stash)
-{
-	int		dotdots;
-	int		start;
-	int		end;
-	int		i;
-	char	*current_path;
-	char	*component;
-	char	*target;
-
-	i = 0;
-	if (!path || !stash || !stash->pwd_backup)
-		return (NULL);
-	dotdots = count_leading_dotdots(path);
-	while(path[i] && (path[i] == '.' || path[i] == '/'))
-		i++;
-	start = i;
-	if (path[start] == '/')
-		start++;
-	current_path = NULL;
-	end = 0;
-	component = get_next_component(path, &start, &end);
-	if (!component && dotdots > 0)
-	{
-		target = build_target_path(dotdots, NULL, stash);
-			return (target);
-	}
-	while (component)
-	{
-		target = build_target_path(dotdots, component, stash);
-		if (!target)
-			return (NULL);
-		if (access(target, F_OK) == 0)
-			current_path = target;
-		else
-			break;
-		start = end;
-		if (!path[start])
-			break;
-		component = get_next_component(path, &start, &end);
-		dotdots = 0;
-	}
-	return (current_path);
 }
